+++ pyoptsparse/pyOpt_constraint.py
@@ -8,8 +8,8 @@ import numpy as np
 
 # Local modules
 from .pyOpt_error import Error, pyOptSparseWarning
+from .pyOpt_types import Dict1DType
 from .pyOpt_utils import INFINITY, _broadcast_to_array, convertToCOO
-from .types import Dict1DType
 
 
 class Constraint:
