+++ pyoptsparse/pyOpt_utils.py
@@ -8,6 +8,7 @@ We use a very simple dictionary format to represent the three most common forms 
     mat = {'csr':[rowp, colind, data], 'shape':[nrow, ncols]} # A csr matrix
     mat = {'csc':[colp, rowind, data], 'shape':[nrow, ncols]} # A csc matrix
 """
+
 # Standard Python modules
 import importlib
 import os
@@ -24,7 +25,7 @@ from scipy.sparse import spmatrix
 
 # Local modules
 from .pyOpt_error import Error
-from .types import ArrayType
+from .pyOpt_types import ArrayType
 
 # Define index mnemonics
 IROW = 0
